rename code to use "modifier"
