+++ src/constants/interfaces/ModifierConfig.ts
@@ -1,12 +1,12 @@
-import GeneratorRequest from "./GeneratorRequest";
+import ModifierRequest from "./ModifierRequest";
 import Param from "./Param";
 
-interface GeneratorConfig {
-  function: (request: GeneratorRequest) => string | Promise<string>;
+interface ModifierConfig {
+  function: (request: ModifierRequest) => string | Promise<string>;
   label: string;
   value: string;
   hidden?: boolean;
   params?: Param[];
 }
 
-export default GeneratorConfig;
+export default ModifierConfig;
