+++ src/modifiers/australia/modifier.ts
@@ -1,12 +1,12 @@
-import GeneratorRequest from "@/constants/interfaces/GeneratorRequest";
+import ModifierRequest from "@/constants/interfaces/ModifierRequest";
 import australiaMap from "../../constants/maps/australia";
-import { getKeyByValue } from "../../globals/generatorHelpers";
+import { getKeyByValue } from "../../globals/modifierHelpers";
 
 const DEFAULT_PARAMS = {
   reverse: false,
 };
 
-export default function generate(request: GeneratorRequest) {
+export default function modify(request: ModifierRequest) {
   const params = { ...DEFAULT_PARAMS, ...request.params };
   let result = "";
   request.text
