+++ src/modifiers/braille/modifier.ts
@@ -1,11 +1,11 @@
-import GeneratorRequest from "@/constants/interfaces/GeneratorRequest";
+import ModifierRequest from "@/constants/interfaces/ModifierRequest";
 import braille from "braille";
 
 const DEFAULT_PARAMS = {
   reverse: false,
 };
 
-export default function generate(request: GeneratorRequest) {
+export default function modify(request: ModifierRequest) {
   const params = { ...DEFAULT_PARAMS, ...request.params };
 
   const str = request.text
