+++ src/modifiers/fancy/modifier.ts
@@ -1,7 +1,7 @@
-import GeneratorRequest from "@/constants/interfaces/GeneratorRequest";
+import ModifierRequest from "@/constants/interfaces/ModifierRequest";
 import fancy from "../../constants/maps/fancy";
 
-export default function generate(request: GeneratorRequest) {
+export default function modify(request: ModifierRequest) {
   let result = "";
   request.text.split("").forEach((c) => {
     result += fancy[c as keyof typeof fancy] ?? c;
