+++ src/modifiers/fraktur/config.ts
@@ -1,8 +1,8 @@
-import GeneratorConfig from "@/constants/interfaces/GeneratorConfig";
+import ModifierConfig from "@/constants/interfaces/ModifierConfig";
 import ParamType from "../../constants/enums/ParamType";
-import generate from "./generator";
+import generate from "./modifier";
 
-const config: GeneratorConfig = {
+const config: ModifierConfig = {
   function: generate,
   label: "𝔉𝔯𝔞𝔨𝔱𝔲𝔯",
   value: "fraktur",
