+++ src/modifiers/fraktur/modifier.ts
@@ -1,11 +1,11 @@
-import GeneratorRequest from "@/constants/interfaces/GeneratorRequest";
+import ModifierRequest from "@/constants/interfaces/ModifierRequest";
 import { fraktur, frakturBold } from "../../constants/maps/fraktur";
 
 const DEFAULT_PARAMS = {
   bold: false,
 };
 
-export default function generate(request: GeneratorRequest) {
+export default function modify(request: ModifierRequest) {
   const params = { ...DEFAULT_PARAMS, ...request.params };
 
   let result = "";
