+++ src/modifiers/leet/modifier.ts
@@ -1,4 +1,4 @@
-import GeneratorRequest from "@/constants/interfaces/GeneratorRequest";
+import ModifierRequest from "@/constants/interfaces/ModifierRequest";
 import {
   alphabetBasic,
   alphabetAdvanced,
@@ -10,7 +10,7 @@ const DEFAULT_PARAMS = {
   reverse: false,
 };
 
-export default function generate(request: GeneratorRequest) {
+export default function modify(request: ModifierRequest) {
   const params = { ...DEFAULT_PARAMS, ...request.params };
 
   let str = request.text;
