+++ src/modifiers/zalgo/config.ts
@@ -1,9 +1,9 @@
-import GeneratorConfig from "@/constants/interfaces/GeneratorConfig";
+import ModifierConfig from "@/constants/interfaces/ModifierConfig";
 import ParamType from "../../constants/enums/ParamType";
-import generate from "./generator";
+import modify from "./modifier";
 
-const config: GeneratorConfig = {
-  function: generate,
+const config: ModifierConfig = {
+  function: modify,
   label: "Z̙̐̏ḁ̇̊l̛̝͑g̢͓̰ơ̜̠",
   value: "zalgo",
   params: [
